+++ vsc-organize-imports/out/OrganizeImports.js
@@ -14,7 +14,7 @@ class OrganizeImports {
     getConfig(property, defaultValue) {
         return vsc.getConfig('vscOrganizeImports', property, defaultValue);
     }
-    run(uri) {
+    run(options, uri) {
         return __awaiter(this, void 0, void 0, function* () {
             if (!uri) {
                 return;
@@ -24,20 +24,12 @@ class OrganizeImports {
                 return;
             }
             const path = uri.path;
-            //load settings:
-            const orderSpecifiers = this.getConfig('orderSpecifiers', false);
-            const orderSpecifiersAsSingleLine = this.getConfig('orderSpecifiersAsSingleLine', false);
-            const baseUrl = this.getConfig('baseUrl', 'src');
-            const emptyLinesAfterGlobalImports = this.getConfig('emptyLinesAfterGlobalImports', 0);
-            const emptyLinesAfterAbsoluteImports = this.getConfig('emptyLinesAfterAbsoluteImports', 0);
-            const emptyLinesLocalImports = this.getConfig('emptyLinesLocalImports', 0);
-            const emptyLinesAfterImports = this.getConfig('emptyLinesAfterImports', 1);
             const rootPath = vsc.getRootPath(path);
             if (!rootPath) {
                 return;
             }
-            const basePath = vsc.joinPaths(rootPath, baseUrl);
-            yield SortImports_1.SortImports(basePath, content, emptyLinesAfterGlobalImports, emptyLinesAfterAbsoluteImports, emptyLinesLocalImports, emptyLinesAfterImports, orderSpecifiers, orderSpecifiersAsSingleLine);
+            options.basePath = vsc.joinPaths(rootPath, options.baseUrl);
+            yield SortImports_1.SortImports(path, content, options);
         });
     }
 }