+++ vsc-organize-imports/src/SortImports.ts
@@ -2,24 +2,46 @@ import * as ts from 'typescript'
 import * as vsc from 'vsc-base'
 import * as vscode from 'vscode'
 
-type Imports = {
+
+type ImportSpecifier = { fullString: string, name: string, node: ts.ImportSpecifier }
+
+type Import = {
   node: ts.ImportDeclaration
   pos: vsc.VscodePosition
   name?: string
-  specifiers: { fullString: string, name: string, node: ts.ImportSpecifier }[]
+  sortName: string
+  specifiers: ImportSpecifier[]
   path: string
   fullString: string
-}[]
+}
 
-export async function SortImports(
+type Imports = Import[]
+
+type groupType =
+  //local (absolute)
+  "absoluteDirect" |
+  "absolute" |
+  //local (relative)
+  "relativeDirect" |
+  "relative" |
+  //globals
+  "globalDirect" |
+  "global"
+
+export type SortImportsOptions = {
+  baseUrl: string,
   basePath: string,
-  content: string,
-  emptyLinesAfterGlobalImports: number,
-  emptyLinesAfterAbsoluteImports: number,
-  emptyLinesLocalImports: number,
-  emptyLinesAfterImports: number,
   orderSpecifiers: boolean,
   orderSpecifiersAsSingleLine: boolean
+  emptyLinesAfterImports: number
+  emptyLinesBetweenFilledGroups: number
+  groups: { groups: groupType[], emptyLines: true, sortBy: string }[]
+}
+
+export async function SortImports(
+  path: string,
+  content: string,
+  options: SortImportsOptions
 ): Promise<vscode.TextEdit[] | undefined> {
 
   // Find first non imports: (exclude 'use strict' and sourceFile')
@@ -38,7 +60,7 @@ export async function SortImports(
     ts.isImportDeclaration(node) && (!firstNode || node.pos < firstNode.pos)) as ts.ImportDeclaration[]
 
   //Find first node that is not in import
-  const imports = mapImports(content, _imports);
+  const imports = mapImports(content, _imports, options);
   if (!imports) {
     return Promise.resolve(undefined)
   }
@@ -46,17 +68,12 @@ export async function SortImports(
   const firstImport = imports[0]
   const lastImport = imports[imports.length - 1]
 
-  if (orderSpecifiers) {
-    sortNamedImports(imports, orderSpecifiersAsSingleLine)
-  }
+  const fillDir = vsc.getDir(path)
 
   const newImportContent = await organizeImports(
+    fillDir,
     imports,
-    basePath,
-    emptyLinesAfterGlobalImports,
-    emptyLinesAfterAbsoluteImports,
-    emptyLinesLocalImports,
-    emptyLinesAfterImports
+    options
   )
 
   let end = lastImport.node.end
@@ -66,112 +83,133 @@ export async function SortImports(
 
 }
 
-const sortNamedImports = (imports: Imports, orderSpecifiersAsSingleLine: boolean) => {
-  imports.map(imp => {
-    if (imp.specifiers) {
-      imp.specifiers.sort((a, b) => a.name.localeCompare(b.name))
-      if (orderSpecifiersAsSingleLine) {
-        const specifierContent = imp.specifiers.map(s => s.fullString).join(', ')
-        imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{ ' + specifierContent + ' }')
-      } else {
-        const specifierContent = imp.specifiers.map(s => s.fullString).join(',\n  ')
-        imp.fullString = imp.fullString.replace(/\{[^}]+\}/, '{\n  ' + specifierContent + '\n}')
-      }
-    }
-  })
+const sortNamedImports = (specifiers: ImportSpecifier[], fullString: string, orderSpecifiersAsSingleLine: boolean) => {
+  specifiers.sort((a, b) => a.name.localeCompare(b.name))
+  if (orderSpecifiersAsSingleLine) {
+    const specifierContent = specifiers.map(s => s.fullString).join(', ')
+    fullString = fullString.replace(/\{[^}]+\}/, '{ ' + specifierContent + ' }')
+  } else {
+    const specifierContent = specifiers.map(s => s.fullString).join(',\n  ')
+    fullString = fullString.replace(/\{[^}]+\}/, '{\n  ' + specifierContent + '\n}')
+  }
+  return fullString;
 }
 
 const organizeImports = async (
+  fillDir: string,
   imports: Imports,
-  basePath: string,
-  emptyLinesAfterGlobalImports: number,
-  emptyLinesAfterAbsoluteImports: number,
-  emptyLinesLocalImports: number,
-  emptyLinesAfterImports: number,
+  options: SortImportsOptions
 ) => {
-  const moduleExtensionRegExp = /\.([tj]sx?)+$/;
-  const localRegExp = /^\./;
-
+  const relativeRegExp = /^\./;
   const groups = {
-    localImports: [] as Imports,
-    absoluteImports: [] as Imports,
-    directImports: [] as Imports,
-    globalImports: [] as Imports,
+    //locals (absolute)
+    absoluteDirect: [] as Imports,
+    absolute: [] as Imports,
+    //local (relative)
+    relativeDirect: [] as Imports,
+    relative: [] as Imports,
+    //globals
+    globalDirect: [] as Imports,
+    global: [] as Imports,
   }
   //split into global / local
   imports.forEach(_import => {
-    if (localRegExp.test(_import.path)) {
-      groups.localImports.push(_import)
-      return
+    let fullPath = '';
+    const relative = relativeRegExp.test(_import.path)
+    if (relative) {
+      fullPath = vsc.joinPaths(fillDir, _import.path)
+    } else {
+      fullPath = vsc.joinPaths(options.basePath, _import.path)
     }
-    const fullPath = vsc.joinPaths(basePath, _import.path)
-    if (moduleExtensionRegExp.test(_import.path) && vsc.doesExists(fullPath)) {
-      groups.absoluteImports.push(_import)
-      return
-    }
+
+    //base groups settings:
+    let local = vsc.doesExists(fullPath)
     if (
+      !local &&
       vsc.doesExists(fullPath + '.ts') ||
       vsc.doesExists(fullPath + '.tsx') ||
       vsc.doesExists(fullPath + '.js') ||
       vsc.doesExists(fullPath + '.jsx')
     ) {
-      groups.absoluteImports.push(_import)
-      return
+      local = true
     }
-    if (!_import.node.importClause) {
-      groups.directImports.push(_import)
-      return
+    const direct = !_import.node.importClause
+    const hasDefault = !!_import.name
+    const hasNamed = _import.specifiers.length > 0
+
+    if (local && relative) {
+      if (direct) {
+        groups.relativeDirect.push(_import)
+      } else {
+        groups.relative.push(_import)
+      }
+    } else if (local) {
+      if (direct) {
+        groups.absoluteDirect.push(_import)
+      } else {
+        groups.absolute.push(_import)
+      }
+    } else {
+      if (direct) {
+        groups.globalDirect.push(_import)
+      } else if (hasDefault && hasNamed) {
+        groups.global.push(_import)
+      }
     }
-    groups.globalImports.push(_import)
   })
-  const defaultMapping: { groups: (keyof typeof groups)[], emptyLinesAfterGroup: number }[] = [
-    { groups: ['globalImports'], emptyLinesAfterGroup: emptyLinesAfterGlobalImports },
-    { groups: ['absoluteImports'], emptyLinesAfterGroup: emptyLinesAfterAbsoluteImports },
-    { groups: ['localImports'], emptyLinesAfterGroup: emptyLinesLocalImports },
-    { groups: ['directImports'], emptyLinesAfterGroup: 0 },
-  ]
+  const defaultMapping: { groups: (keyof typeof groups)[], emptyLines: boolean, sortBy: string }[] = options.groups
   let newImportContent = ""
   defaultMapping.forEach((groupOptions, index) => {
     let group = [] as Imports
     groupOptions.groups.forEach(groupName => {
-      group = [...groups[groupName]]
+      group = [...group, ...groups[groupName]]
     })
     if (group.length === 0) {
       return
     }
     // sort
-    group.sort((a, b) => a.path.localeCompare(b.path))
+    if (groupOptions.sortBy === 'path') {
+      group.sort((a, b) => a.path.localeCompare(b.path))
+    } else if (groupOptions.sortBy === 'name') {
+      group.sort((a, b) => a.sortName.localeCompare(b.sortName))
+    } else {
+      group.sort((a, b) => a.node.getText().localeCompare(b.node.getText()))
+    }
+
     // join and add
     newImportContent += group.map(imp => imp.fullString).join('\n') + '\n'
-    // add spaces
-    for (let space = 0; space < groupOptions.emptyLinesAfterGroup; space++) {
-      newImportContent += '\n';
+    if (!newImportContent.match(/\n\n$/) && groupOptions.emptyLines) {
+      // add spaces
+      for (let space = 0; space < options.emptyLinesBetweenFilledGroups; space++) {
+        newImportContent = '\n' + newImportContent;
+      }
     }
   })
   newImportContent = newImportContent.trim() + '\n'
-  for (let lines = 0; lines < emptyLinesAfterImports; lines++) {
+  for (let lines = 0; lines < options.emptyLinesAfterImports; lines++) {
     newImportContent += '\n';
   }
   return newImportContent
 }
 
-const mapImports = (content: string, _imports: ts.ImportDeclaration[]) => {
+const mapImports = (content: string, _imports: ts.ImportDeclaration[], options: SortImportsOptions) => {
   // All imports before first statement, mapped with import path
   // Map with name?, fullString, and named imports info
   const imports: Imports = _imports.map((node, index) => {
-    let name = ''
-    const fullString = content
+    let name = '', sortName = ''
+    let importFullString = content
       .substring(
         index === 0 ? node.pos : _imports[index - 1].end + 1,
         node.end
       )
       .trim();
-    let specifiers: { fullString: string, name: string, node: ts.ImportSpecifier }[] = []
+    let specifiers: ImportSpecifier[] = []
     const importClause = node.importClause
     //named imports (specifiers)
     if (importClause) {
       if (importClause.name) {
         name = importClause.name.getText()
+        sortName = name
       }
       if (importClause.namedBindings && ts.isNamedImports(importClause.namedBindings)) {
         specifiers = importClause.namedBindings.elements.map(e => ({
@@ -179,14 +217,19 @@ const mapImports = (content: string, _imports: ts.ImportDeclaration[]) => {
           node: e,
           name: e.name.getText()
         }))
+        if (options.orderSpecifiers) {
+          importFullString = sortNamedImports(specifiers, importFullString, options.orderSpecifiersAsSingleLine)
+        }
+        sortName = sortName + specifiers.map(s => s.name).join()
       }
     }
     const pos = vsc.createVscodeRangeAndPosition(content, node.pos, node.end);
     return ({
       name,
+      sortName,
       pos,
       specifiers,
-      fullString,
+      fullString: importFullString,
       node: node,
       path: node.moduleSpecifier
         .getText()