+++ vsc-organize-imports/src/extension.ts
@@ -1,21 +1,23 @@
 import * as vscode from 'vscode'
 import OrganizeImports from './OrganizeImports'
+import { SortImportsOptions } from './SortImports';
+import * as vsc from 'vsc-base'
+
+const cleanCode = new OrganizeImports()
 
 // The module 'vscode' contains the VS Code extensibility API
 
 // Import the module and reference it with the alias vscode in your code below
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
-export function activate(context: vscode.ExtensionContext) {
+export async function activate(context: vscode.ExtensionContext) {
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.log('Congratulations, your extension "vsc-script" is now active!')
+	console.log('Congratulations, your extension "vsc-organize-imports" is now active!')
 
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-
-	const cleanCode = new OrganizeImports()
 
 	let disposableShortcut = vscode.commands.registerTextEditorCommand(
 		'extension.vscOrganizeImports',
@@ -23,7 +25,7 @@ export function activate(context: vscode.ExtensionContext) {
 			if (!vscode.window.activeTextEditor) {
 				return
 			}
-			cleanCode.run(vscode.window.activeTextEditor.document.uri)
+			runExtension(vscode.window.activeTextEditor.document.uri)
 		}
 	)
 
@@ -31,7 +33,7 @@ export function activate(context: vscode.ExtensionContext) {
 		const formatOnSave = cleanCode.getConfig('formatOnSave', true)
 		if (formatOnSave) {
 			event.waitUntil(
-				cleanCode.run(event.document.uri)
+				runExtension(event.document.uri)
 			);
 		}
 	})
@@ -45,3 +47,32 @@ export function activate(context: vscode.ExtensionContext) {
 export function deactivate() { }
 
 
+
+const runExtension = async (uri?: vscode.Uri) => {
+	if (!uri) {
+		return;
+	}
+	const rootPath = vsc.getRootPath(uri.fsPath);
+	if (!rootPath) {
+		return
+	}
+	const options = await getOptions(rootPath)
+	if (!options) {
+		return
+	}
+	return cleanCode.run(options, uri)
+}
+
+
+const getOptions = async (rootPath: string) => {
+	//load settings:
+	const packageJson = await vsc.getRootPackageJson<{
+		"vsc-organize-imports": SortImportsOptions
+	}>(rootPath);
+	if (packageJson["vsc-organize-imports"] === undefined) {
+		vsc.showErrorMessage("vsc-organize-imports was not found in the package.json root file!")
+		return;
+	}
+	const packageJsonOptions = packageJson["vsc-organize-imports"] as SortImportsOptions;
+	return packageJsonOptions;
+}