+++ vsc-script/src/Script.ts
@@ -1,11 +1,11 @@
 'use strict'
 import * as cp from 'child-process-promise'
 import * as fs from 'fs-extra'
-import * as ts from 'typescript'
-import * as vsc from 'vsc-base'
-import * as vscode from 'vscode'
 import * as path from 'path'
-//import * as vsc from './vsc-base-development/vsc-base'
+import * as ts from 'typescript'
+//import * as vsc from 'vsc-base'
+import * as vscode from 'vscode'
+import * as vsc from './vsc-base-development/vsc-base'
 
 export default class Script {
    /**